pyproject.toml
[tool.poetry]
name = "wandregal"
version = "0.1.0"
description = ""
authors = ["Sebastian Herold <sebastian.herold@zalando.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pycollada = "^0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
